+++ apps/ungdomsytelse-deltaker/src/analytics/analytics.tsx
@@ -28,7 +28,6 @@ export enum ApiError {
 
 interface Props {
     applicationKey: string;
-    logToConsoleOnly?: boolean;
     isActive?: boolean;
     children: React.ReactNode;
     maxAwaitTime?: number;
@@ -55,7 +54,7 @@ export const registerAnalytics = (websiteId?: string) => {
 };
 
 export const [AnalyticsProvider, useAnalyticsInstance] = constate((props: Props) => {
-    const { applicationKey, isActive = true, maxAwaitTime = MAX_AWAIT_TIME, logToConsoleOnly } = props;
+    const { applicationKey, isActive = true, maxAwaitTime = MAX_AWAIT_TIME } = props;
 
     async function logEvent(eventName: string, eventProperties?: EventProperties) {
         const logger = getAnalyticsInstance('dekoratoren');
@@ -63,15 +62,9 @@ export const [AnalyticsProvider, useAnalyticsInstance] = constate((props: Props)
             const timeoutPromise = new Promise((resolve) => setTimeout(() => resolve(null), maxAwaitTime));
             const logPromise = new Promise((resolve) => {
                 const eventProps = { ...eventProperties, applikasjon: applicationKey };
-                if (logToConsoleOnly) {
-                    // eslint-disable-next-line no-console
-                    console.log({ eventName, eventProperties: eventProps });
+                logger(eventName, eventProps).catch(() => {
                     resolve(true);
-                } else {
-                    logger(eventName, eventProps).catch(() => {
-                        resolve(true);
-                    });
-                }
+                });
             });
             return Promise.race([timeoutPromise, logPromise]);
         }