+++ apps/ungdomsytelse-deltaker/src/pages/HentDeltakerErrorPage.tsx
@@ -1,6 +1,6 @@
 import { Alert, BodyShort, Heading } from '@navikt/ds-react';
 import SifGuidePanel from '@navikt/sif-common-core-ds/src/components/sif-guide-panel/SifGuidePanel';
-import { ErrorPage } from '@navikt/sif-common-soknad-ds/src';
+import Feilside from './Feilside';
 
 interface Props {
     error: string;
@@ -8,22 +8,16 @@ interface Props {
 
 const HentDeltakerErrorPage = ({ error }: Props) => {
     return (
-        <ErrorPage
-            pageTitle="Det oppstod en feil"
-            bannerTitle="Søknad om ungdomsprogramytelse"
-            contentRenderer={() => {
-                return (
-                    <SifGuidePanel mood="uncertain">
-                        <Heading level="2" size="medium" spacing>
-                            Det oppstod en feil under henting av informasjon
-                        </Heading>
-                        <Alert variant="error" inline>
-                            <BodyShort>{error}</BodyShort>
-                        </Alert>
-                    </SifGuidePanel>
-                );
-            }}
-        />
+        <Feilside pageTitle="Det oppstod en feil" bannerTitle="Søknad om ungdomsprogramytelse">
+            <SifGuidePanel mood="uncertain">
+                <Heading level="2" size="medium" spacing>
+                    Det oppstod en feil under henting av informasjon
+                </Heading>
+                <Alert variant="error" inline>
+                    <BodyShort>{error}</BodyShort>
+                </Alert>
+            </SifGuidePanel>
+        </Feilside>
     );
 };